refactor(reddit): move validation into another
